feat: merged oidc and saml into one loader
