+++ packages/cli/src/instance-settings-loader/__tests__/instance-settings-loader.service.test.ts
@@ -2,66 +2,54 @@ import { mock } from 'jest-mock-extended';
 import type { Logger } from '@n8n/backend-common';
 
 import { InstanceSettingsLoaderService } from '../instance-settings-loader.service';
-import type { OidcInstanceSettingsLoader } from '../loaders/oidc.instance-settings-loader';
 import type { OwnerInstanceSettingsLoader } from '../loaders/owner.instance-settings-loader';
-import type { SamlInstanceSettingsLoader } from '../loaders/saml.instance-settings-loader';
 import type { SecurityPolicyInstanceSettingsLoader } from '../loaders/security-policy.instance-settings-loader';
+import type { SsoInstanceSettingsLoader } from '../loaders/sso.instance-settings-loader';
 
 describe('InstanceSettingsLoaderService', () => {
 	const logger = mock<Logger>({ scoped: jest.fn().mockReturnThis() });
 	const ownerLoader = mock<OwnerInstanceSettingsLoader>();
-	const oidcLoader = mock<OidcInstanceSettingsLoader>();
+	const ssoLoader = mock<SsoInstanceSettingsLoader>();
 	const securityPolicyLoader = mock<SecurityPolicyInstanceSettingsLoader>();
-	const samlLoader = mock<SamlInstanceSettingsLoader>();
 
 	beforeEach(() => {
 		jest.resetAllMocks();
 		logger.scoped.mockReturnThis();
 		ownerLoader.run.mockResolvedValue('skipped');
-		oidcLoader.run.mockResolvedValue('skipped');
+		ssoLoader.run.mockResolvedValue('skipped');
 		securityPolicyLoader.run.mockResolvedValue('skipped');
-		samlLoader.run.mockResolvedValue('skipped');
 	});
 
 	const createService = () =>
-		new InstanceSettingsLoaderService(
-			logger,
-			ownerLoader,
-			oidcLoader,
-			securityPolicyLoader,
-			samlLoader,
-		);
+		new InstanceSettingsLoaderService(logger, ownerLoader, ssoLoader, securityPolicyLoader);
 
 	it('should run all loaders', async () => {
 		await createService().init();
 
 		expect(ownerLoader.run).toHaveBeenCalled();
-		expect(oidcLoader.run).toHaveBeenCalled();
+		expect(ssoLoader.run).toHaveBeenCalled();
 		expect(securityPolicyLoader.run).toHaveBeenCalled();
-		expect(samlLoader.run).toHaveBeenCalled();
 	});
 
 	it('should log each loader result', async () => {
 		ownerLoader.run.mockResolvedValue('created');
-		samlLoader.run.mockResolvedValue('created');
+		ssoLoader.run.mockResolvedValue('created');
 
 		await createService().init();
 
 		expect(logger.debug).toHaveBeenCalledWith('Instance settings loader "owner": created');
-		expect(logger.debug).toHaveBeenCalledWith('Instance settings loader "oidc": skipped');
+		expect(logger.debug).toHaveBeenCalledWith('Instance settings loader "sso": created');
 		expect(logger.debug).toHaveBeenCalledWith(
 			'Instance settings loader "security-policy": skipped',
 		);
-		expect(logger.debug).toHaveBeenCalledWith('Instance settings loader "saml": created');
 	});
 
 	it('should stop execution if a loader throws', async () => {
-		oidcLoader.run.mockRejectedValue(new Error('oidc failure'));
+		ssoLoader.run.mockRejectedValue(new Error('sso failure'));
 
-		await expect(createService().init()).rejects.toThrow('oidc failure');
+		await expect(createService().init()).rejects.toThrow('sso failure');
 
 		expect(ownerLoader.run).toHaveBeenCalled();
 		expect(securityPolicyLoader.run).not.toHaveBeenCalled();
-		expect(samlLoader.run).not.toHaveBeenCalled();
 	});
 });
