+++ packages/cli/src/instance-settings-loader/instance-settings-loader.service.ts
@@ -1,10 +1,9 @@
 import { Logger } from '@n8n/backend-common';
 import { Service } from '@n8n/di';
 
-import { OidcInstanceSettingsLoader } from './loaders/oidc.instance-settings-loader';
 import { OwnerInstanceSettingsLoader } from './loaders/owner.instance-settings-loader';
-import { SamlInstanceSettingsLoader } from './loaders/saml.instance-settings-loader';
 import { SecurityPolicyInstanceSettingsLoader } from './loaders/security-policy.instance-settings-loader';
+import { SsoInstanceSettingsLoader } from './loaders/sso.instance-settings-loader';
 
 type LoaderResult = 'created' | 'skipped';
 
@@ -13,18 +12,16 @@ export class InstanceSettingsLoaderService {
 	constructor(
 		private logger: Logger,
 		private readonly ownerLoader: OwnerInstanceSettingsLoader,
-		private readonly oidcLoader: OidcInstanceSettingsLoader,
+		private readonly ssoLoader: SsoInstanceSettingsLoader,
 		private readonly securityPolicyLoader: SecurityPolicyInstanceSettingsLoader,
-		private readonly samlLoader: SamlInstanceSettingsLoader,
 	) {
 		this.logger = this.logger.scoped('instance-settings-loader');
 	}
 
 	async init(): Promise<void> {
 		await this.run('owner', async () => await this.ownerLoader.run());
-		await this.run('oidc', async () => await this.oidcLoader.run());
+		await this.run('sso', async () => await this.ssoLoader.run());
 		await this.run('security-policy', async () => await this.securityPolicyLoader.run());
-		await this.run('saml', async () => await this.samlLoader.run());
 	}
 
 	private async run(name: string, fn: () => Promise<LoaderResult>): Promise<void> {
