+++ packages/cli/test/integration/oidc/oidc.instance-settings-loader.test.ts
@@ -3,10 +3,11 @@ import { GlobalConfig } from '@n8n/config';
 import { SettingsRepository } from '@n8n/db';
 import { Container } from '@n8n/di';
 
-import { OidcInstanceSettingsLoader } from '@/instance-settings-loader/loaders/oidc.instance-settings-loader';
+import { SsoInstanceSettingsLoader } from '@/instance-settings-loader/loaders/sso.instance-settings-loader';
 import { PROVISIONING_PREFERENCES_DB_KEY } from '@/modules/provisioning.ee/constants';
 import { OIDC_PREFERENCES_DB_KEY } from '@/modules/sso-oidc/constants';
 import { OidcService } from '@/modules/sso-oidc/oidc.service.ee';
+import { SAML_PREFERENCES_DB_KEY } from '@/modules/sso-saml/constants';
 
 beforeAll(async () => {
 	await testDb.init();
@@ -16,7 +17,7 @@ afterAll(async () => {
 	await testDb.terminate();
 });
 
-describe('OidcInstanceSettingsLoader → OidcService roundtrip', () => {
+describe('SsoInstanceSettingsLoader → OidcService roundtrip', () => {
 	let originalConfig: Record<string, unknown>;
 
 	beforeEach(() => {
@@ -33,6 +34,7 @@ describe('OidcInstanceSettingsLoader → OidcService roundtrip', () => {
 		// Clean up DB rows
 		const settingsRepository = Container.get(SettingsRepository);
 		await settingsRepository.delete({ key: OIDC_PREFERENCES_DB_KEY });
+		await settingsRepository.delete({ key: SAML_PREFERENCES_DB_KEY });
 		await settingsRepository.delete({ key: PROVISIONING_PREFERENCES_DB_KEY });
 	});
 
@@ -49,7 +51,7 @@ describe('OidcInstanceSettingsLoader → OidcService roundtrip', () => {
 			ssoUserRoleProvisioning: 'instance_and_project_roles',
 		});
 
-		const loader = Container.get(OidcInstanceSettingsLoader);
+		const loader = Container.get(SsoInstanceSettingsLoader);
 		await loader.run();
 
 		const oidcService = Container.get(OidcService);
